requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
holidays>=0.34
fpdf2>=2.7.0
